Share the common editor code
